added parameters section to project manager and removed from task manager, task designer unchanged and set defaults not implemeneted in project manager
