+++ squidApp/src/main/java/org/cirdles/squid/gui/TaskManagerController.java
@@ -15,48 +15,28 @@
  */
 package org.cirdles.squid.gui;
 
-import java.net.URL;
-import java.util.ResourceBundle;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.CheckBox;
-import javafx.scene.control.ComboBox;
-import javafx.scene.control.Label;
-import javafx.scene.control.RadioButton;
-import javafx.scene.control.Spinner;
-import javafx.scene.control.SpinnerValueFactory;
-import javafx.scene.control.TextArea;
-import javafx.scene.control.TextField;
-import javafx.scene.control.ToggleGroup;
+import javafx.scene.control.*;
 import javafx.scene.layout.GridPane;
 import javafx.util.StringConverter;
 import org.cirdles.squid.constants.Squid3Constants;
 import org.cirdles.squid.constants.Squid3Constants.TaskTypeEnum;
-
-import static org.cirdles.squid.gui.SquidUIController.squidLabData;
-import static org.cirdles.squid.gui.SquidUIController.squidProject;
-import static org.cirdles.squid.gui.constants.Squid3GuiConstants.STYLE_MANAGER_TITLE;
-
-import org.cirdles.squid.parameters.parameterModels.ParametersModel;
 import org.cirdles.squid.tasks.TaskInterface;
 import org.cirdles.squid.tasks.expressions.Expression;
-import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.PARENT_ELEMENT_CONC_CONST;
-import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.PARENT_ELEMENT_CONC_CONST_DEFAULT_EXPRESSION;
-import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.TH_U_EXP_DEFAULT;
-import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.TH_U_EXP_DEFAULT_EXPRESSION;
-import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.UNCOR206PB238U_CALIB_CONST;
-import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.UNCOR208PB232TH_CALIB_CONST;
-import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.TH_U_EXP_RM;
-import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.UNCOR206PB238U_CALIB_CONST_DEFAULT_EXPRESSION;
-import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.UNCOR208PB232TH_CALIB_CONST_DEFAULT_EXPRESSION;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
 import static org.cirdles.squid.gui.SquidUI.HEALTHY_EXPRESSION_STYLE;
 import static org.cirdles.squid.gui.SquidUI.UNHEALTHY_EXPRESSION_STYLE;
+import static org.cirdles.squid.gui.SquidUIController.squidProject;
+import static org.cirdles.squid.gui.constants.Squid3GuiConstants.STYLE_MANAGER_TITLE;
 import static org.cirdles.squid.tasks.expressions.Expression.makeExpressionForAudit;
+import static org.cirdles.squid.tasks.expressions.builtinExpressions.BuiltInExpressionsDataDictionary.*;
 import static org.cirdles.squid.utilities.conversionUtilities.RoundingUtilities.useSigFig15;
 
 /**
@@ -67,6 +47,9 @@ import static org.cirdles.squid.utilities.conversionUtilities.RoundingUtilities.
 public class TaskManagerController implements Initializable {
 
     private TaskInterface task;
+
+    @FXML
+    private TextField provenanceTextField;
     @FXML
     private GridPane taskManagerGridPane;
     @FXML
@@ -78,43 +61,15 @@ public class TaskManagerController implements Initializable {
     @FXML
     private RadioButton generalTaskTypeRadioButton;
     @FXML
-    private ToggleGroup toggleGroupSMB;
-    @FXML
-    private ToggleGroup toggleGroupRatioCalcMethod;
-    @FXML
     private TextField authorsNameTextField;
     @FXML
     private TextField labNameTextField;
     @FXML
     private TextField taskDescriptionTextField;
     @FXML
-    private TextField provenanceTextField;
-    @FXML
-    private RadioButton yesSBMRadioButton;
-    @FXML
-    private RadioButton noSBMRadioButton;
-    @FXML
-    private RadioButton linearRegressionRatioCalcRadioButton;
-    @FXML
-    private RadioButton spotAverageRatioCalcRadioButton;
-    @FXML
     private TextArea taskAuditTextArea;
     @FXML
-    private RadioButton pb204RadioButton;
-    @FXML
-    private ToggleGroup toggleGroupIsotope;
-    @FXML
-    private RadioButton pb207RadioButton;
-    @FXML
-    private RadioButton pb208RadioButton;
-    @FXML
     private Label titleLabel;
-    @FXML
-    private CheckBox autoExcludeSpotsCheckBox;
-    @FXML
-    private ComboBox<ParametersModel> commonPbModelComboBox;
-    @FXML
-    private ComboBox<ParametersModel> physConstModelComboBox;
     @FXML
     private ToggleGroup primaryAgeToggleGroup;
     @FXML
@@ -135,18 +90,6 @@ public class TaskManagerController implements Initializable {
     private Label parentConcExpressionLabel;
     @FXML
     private Label parentConcLabel;
-    @FXML
-    private Spinner<Double> assignedExternalErrUSpinner;
-    @FXML
-    private Spinner<Double> assignedExternalErrThSpinner;
-    @FXML
-    private ComboBox<String> delimiterComboBox;
-    @FXML
-    private RadioButton roundingSquid25;
-    @FXML
-    private ToggleGroup roundingToggleGroup;
-    @FXML
-    private RadioButton roundingSquid3;
 
     /**
      * Initializes the controller class.
@@ -158,12 +101,10 @@ public class TaskManagerController implements Initializable {
     public void initialize(URL url, ResourceBundle rb) {
         if (squidProject.getTask() != null) {
             task = squidProject.getTask();
-            roundingSquid3.setSelected(task.isRoundingForSquid3());
             useSigFig15 = task.isRoundingForSquid3();
             task.setupSquidSessionSpecsAndReduceAndReport(false);
             populateTaskFields();
             setupListeners();
-            setUpParametersModelsComboBoxes();
         } else {
             taskAuditTextArea.setText("No Task information available");
         }
@@ -185,84 +126,11 @@ public class TaskManagerController implements Initializable {
             geochronTaskTypeRadioButton.setSelected(true);
         }
 
-        if (task.isUseSBM()) {
-            yesSBMRadioButton.setSelected(true);
-        } else {
-            noSBMRadioButton.setSelected(true);
-        }
-
-        if (task.isUserLinFits()) {
-            linearRegressionRatioCalcRadioButton.setSelected(true);
-        } else {
-            spotAverageRatioCalcRadioButton.setSelected(true);
-        }
-
         taskAuditTextArea.setText(task.printTaskAudit());
-
-        // set Pb208 Isotope selector visible or not
-        pb208RadioButton.setVisible(!task.isDirectAltPD() && !task.getParentNuclide().contains("232"));
-
-        switch (task.getSelectedIndexIsotope()) {
-            case PB_204:
-                pb204RadioButton.setSelected(true);
-                break;
-            case PB_207:
-                pb207RadioButton.setSelected(true);
-                break;
-            case PB_208:
-                pb208RadioButton.setSelected(true);
-                break;
-        }
-
-        autoExcludeSpotsCheckBox.setSelected(task.isSquidAllowsAutoExclusionOfSpots());
-
-        SpinnerValueFactory<Double> valueFactoryU
-                = new SpinnerValueFactory.DoubleSpinnerValueFactory(0.50, 1.00, task.getExtPErrU(), 0.05);
-        assignedExternalErrUSpinner.setValueFactory(valueFactoryU);
-        assignedExternalErrUSpinner.valueProperty().addListener((ObservableValue<? extends Double> observable, 
-                Double oldValue, Double newValue) -> {
-            task.setExtPErrU(newValue);
-        });
-
-        SpinnerValueFactory<Double> valueFactoryTh
-                = new SpinnerValueFactory.DoubleSpinnerValueFactory(0.50, 1.00, task.getExtPErrTh(), 0.05);
-        assignedExternalErrThSpinner.setValueFactory(valueFactoryTh);
-        assignedExternalErrThSpinner.valueProperty().addListener((ObservableValue<? extends Double> observable, //
-                Double oldValue, Double newValue) -> {
-            task.setExtPErrTh(newValue);
-        });
-
-        // samples
-        ObservableList<String> delimitersList = FXCollections.observableArrayList(Squid3Constants.SampleNameDelimitersEnum.names());
-        delimiterComboBox.setItems(delimitersList);
-        // set value before adding listener
-        delimiterComboBox.getSelectionModel().select(task.getDelimiterForUnknownNames());
-        delimiterComboBox.getSelectionModel().selectedItemProperty().addListener(new ChangeListener<String>() {
-            @Override
-            public void changed(ObservableValue<? extends String> ov,
-                    final String oldValue, final String newValue) {
-                task.setDelimiterForUnknownNames(newValue);
-            }
-        });
 
         populateDirectives();
     }
 
-    @FXML
-    private void roundingSquid25Action(ActionEvent event) {
-        useSigFig15 = false;
-        task.setRoundingForSquid3(false);
-        task.setChanged(true);
-        task.setupSquidSessionSpecsAndReduceAndReport(true);
-    }
-
-    @FXML
-    private void roundingSquid3Action(ActionEvent event) {
-        useSigFig15 = true;
-        task.setRoundingForSquid3(true);
-        task.setChanged(true);
-        task.setupSquidSessionSpecsAndReduceAndReport(true);
-    }
 
     class MyConverter extends StringConverter<Double> {
 
@@ -279,7 +147,6 @@ public class TaskManagerController implements Initializable {
     }
 
     /**
-     *
      * @param expressionName
      * @param expressionString
      * @return
@@ -296,10 +163,7 @@ public class TaskManagerController implements Initializable {
         boolean uPicked = ((RadioButton) taskManagerGridPane.lookup("#238")).isSelected();
         boolean directPicked = ((RadioButton) taskManagerGridPane.lookup("#direct")).isSelected();
 
-        pb208RadioButton.setVisible(uPicked && !directPicked);
-        if (!pb208RadioButton.isVisible() && task.getSelectedIndexIsotope().compareTo(Squid3Constants.IndexIsoptopesEnum.PB_208) == 0) {
-            pb204RadioButton.setSelected(true);
-        }
+
         boolean perm1 = uPicked && !directPicked;
         boolean perm2 = uPicked && directPicked;
         boolean perm3 = !uPicked && !directPicked;
@@ -335,8 +199,6 @@ public class TaskManagerController implements Initializable {
         parentConcExpressionLabel.setText(parentPPM_ExpressionString);
         parentConcExpressionLabel.setStyle(parentConcExpressionLabel.getStyle()
                 + (makeExpression(PARENT_ELEMENT_CONC_CONST, parentPPM_ExpressionString).amHealthy() ? HEALTHY_EXPRESSION_STYLE : UNHEALTHY_EXPRESSION_STYLE));
-
-        updateDirectiveButtons();
     }
 
     private void setupListeners() {
@@ -393,46 +255,6 @@ public class TaskManagerController implements Initializable {
         taskAuditTextArea.setText(task.printTaskAudit());
     }
 
-    static class ParameterModelStringConverter extends StringConverter<ParametersModel> {
-
-        @Override
-        public String toString(ParametersModel model) {
-            return model.getModelNameWithVersion();
-        }
-
-        @Override
-        public ParametersModel fromString(String string) {
-            return null;
-        }
-    };
-
-    private void setUpParametersModelsComboBoxes() {
-
-        // PhysicalConstantsModels
-        physConstModelComboBox.setConverter(new ParameterModelStringConverter());
-        physConstModelComboBox.setItems(FXCollections.observableArrayList(squidLabData.getPhysicalConstantsModels()));
-        physConstModelComboBox.getSelectionModel().select(task.getPhysicalConstantsModel());
-
-        physConstModelComboBox.valueProperty()
-                .addListener((ObservableValue<? extends ParametersModel> observable, ParametersModel oldValue, ParametersModel newValue) -> {
-                    task.setPhysicalConstantsModel(newValue);
-                    task.setChanged(true);
-                    task.setupSquidSessionSpecsAndReduceAndReport(false);
-                });
-
-        // CommonPbModels
-        commonPbModelComboBox.setConverter(new ParameterModelStringConverter());
-        commonPbModelComboBox.setItems(FXCollections.observableArrayList(squidLabData.getCommonPbModels()));
-        commonPbModelComboBox.getSelectionModel().select(task.getCommonPbModel());
-
-        commonPbModelComboBox.valueProperty()
-                .addListener((ObservableValue<? extends ParametersModel> observable, ParametersModel oldValue, ParametersModel newValue) -> {
-                    task.setCommonPbModel(newValue);
-                    task.setChanged(true);
-                    task.setupSquidSessionSpecsAndReduceAndReport(false);
-                });
-    }
-
     @FXML
     private void geochronTaskTypeRadioButtonAction(ActionEvent event) {
         task.setTaskType(TaskTypeEnum.valueOf(geochronTaskTypeRadioButton.getId()));
@@ -445,69 +267,5 @@ public class TaskManagerController implements Initializable {
         task.setChanged(true);
     }
 
-    @FXML
-    private void yesSBMRadioButtonAction(ActionEvent event) {
-        task.setUseSBM(true);
-        task.setChanged(true);
-        task.setupSquidSessionSpecsAndReduceAndReport(true);
-    }
-
-    @FXML
-    private void noSBMRadioButtonActions(ActionEvent event) {
-        task.setUseSBM(false);
-        task.setChanged(true);
-        task.setupSquidSessionSpecsAndReduceAndReport(true);
-    }
-
-    @FXML
-    private void linearRegressionRatioCalcRadioButtonAction(ActionEvent event) {
-        task.setUserLinFits(true);
-        task.setChanged(true);
-        task.setupSquidSessionSpecsAndReduceAndReport(true);
-    }
-
-    @FXML
-    private void spotAverageRatioCalcRadioButtonAction(ActionEvent event) {
-        task.setUserLinFits(false);
-        task.setChanged(true);
-        task.setupSquidSessionSpecsAndReduceAndReport(true);
-    }
-
-    @FXML
-    private void pb204RadioButtonAction(ActionEvent event) {
-        updateDirectiveButtons();
-        task.setSelectedIndexIsotope(Squid3Constants.IndexIsoptopesEnum.PB_204);
-        task.setChanged(true);
-    }
-
-    @FXML
-    private void pb207RadioButtonAction(ActionEvent event) {
-        updateDirectiveButtons();
-        task.setSelectedIndexIsotope(Squid3Constants.IndexIsoptopesEnum.PB_207);
-        task.setChanged(true);
-    }
-
-    @FXML
-    private void pb208RadioButtonAction(ActionEvent event) {
-        updateDirectiveButtons();
-        task.setSelectedIndexIsotope(Squid3Constants.IndexIsoptopesEnum.PB_208);
-        task.setChanged(true);
-    }
-
-    private void updateDirectiveButtons() {
-        ((RadioButton) taskManagerGridPane.lookup("#232")).setDisable(pb208RadioButton.isSelected());
-        ((RadioButton) taskManagerGridPane.lookup("#direct")).setDisable(pb208RadioButton.isSelected());
-    }
-
-    @FXML
-    private void autoExcludeSpotsCheckBoxAction(ActionEvent event) {
-        // this will cause weighted mean expressions to be changed with boolean flag
-        task.updateRefMatCalibConstWMeanExpressions(autoExcludeSpotsCheckBox.isSelected());
-    }
-
-    @FXML
-    private void refreshModelsAction(ActionEvent event) {
-        task.refreshParametersFromModels(true, true, false);
-    }
 
 }
