+++ squidApp/src/main/resources/org/cirdles/squid/gui/TaskManager.fxml
@@ -63,86 +63,6 @@
             <Label alignment="CENTER_RIGHT" prefHeight="28.0" prefWidth="109.0" style="-fx-text-fill: #3c77c9; -fx-font-weight: bold;" text="Parameters:" />
          </children>
       </VBox>
-      <GridPane hgap="10.0" GridPane.columnIndex="1" GridPane.rowIndex="6">
-        <columnConstraints>
-          <ColumnConstraints halignment="RIGHT" hgrow="NEVER" maxWidth="210.0" minWidth="210.0" prefWidth="210.0" />
-          <ColumnConstraints hgrow="ALWAYS" minWidth="700.0" prefWidth="700.0" />
-        </columnConstraints>
-        <rowConstraints>
-          <RowConstraints vgrow="NEVER" />
-          <RowConstraints vgrow="NEVER" />
-          <RowConstraints vgrow="NEVER" />
-            <RowConstraints vgrow="NEVER" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
-        </rowConstraints>
-         <children>
-            <Label style="-fx-font-size: 14;" text="Normalise Ion Counts for SBM?" />
-            <Label style="-fx-font-size: 14;" text="Ratio Calculation Method:" GridPane.rowIndex="1" />
-            <Label style="-fx-font-size: 14;" text="Preferred index isotope:" GridPane.rowIndex="3" />
-            <Label layoutX="10.0" layoutY="94.0" style="-fx-font-size: 14;" text="Weighted Means of Ref Mat:" GridPane.rowIndex="4" />
-            <HBox alignment="CENTER_LEFT" spacing="30.0" GridPane.columnIndex="1" GridPane.rowIndex="3">
-               <children>
-                  <RadioButton fx:id="pb204RadioButton" mnemonicParsing="false" onAction="#pb204RadioButtonAction" selected="true" text="204Pb">
-                     <toggleGroup>
-                        <ToggleGroup fx:id="toggleGroupIsotope" />
-                     </toggleGroup>
-                  </RadioButton>
-                  <RadioButton fx:id="pb207RadioButton" mnemonicParsing="false" onAction="#pb207RadioButtonAction" text="207Pb" toggleGroup="$toggleGroupIsotope" />
-                  <RadioButton fx:id="pb208RadioButton" mnemonicParsing="false" onAction="#pb208RadioButtonAction" text="208Pb" toggleGroup="$toggleGroupIsotope" />
-               </children>
-            </HBox>
-            <HBox alignment="CENTER_LEFT" spacing="10.0" GridPane.columnIndex="1">
-               <children>
-                  <RadioButton fx:id="yesSBMRadioButton" layoutX="350.0" mnemonicParsing="false" onAction="#yesSBMRadioButtonAction" selected="true" text="Yes">
-                     <toggleGroup>
-                        <ToggleGroup fx:id="toggleGroupSMB" />
-                     </toggleGroup>
-                  </RadioButton>
-                  <RadioButton fx:id="noSBMRadioButton" mnemonicParsing="false" onAction="#noSBMRadioButtonActions" text="No" toggleGroup="$toggleGroupSMB" />
-                  <RadioButton fx:id="roundingSquid25" mnemonicParsing="false" onAction="#roundingSquid25Action" prefHeight="30.0" prefWidth="153.0" selected="true" text="Squid2.5 rounding" textFill="#f80202" translateX="100.0">
-                     <toggleGroup>
-                        <ToggleGroup fx:id="roundingToggleGroup" />
-                     </toggleGroup>
-                  </RadioButton>
-                  <RadioButton fx:id="roundingSquid3" mnemonicParsing="false" onAction="#roundingSquid3Action" prefHeight="29.0" prefWidth="156.0" text="Squid3 rounding" textFill="#f50303" toggleGroup="$roundingToggleGroup" translateX="100.0" />
-               </children>
-            </HBox>
-            <HBox alignment="CENTER_LEFT" spacing="10.0" GridPane.columnIndex="1" GridPane.rowIndex="1">
-               <children>
-                  <RadioButton fx:id="linearRegressionRatioCalcRadioButton" layoutX="350.0" mnemonicParsing="false" onAction="#linearRegressionRatioCalcRadioButtonAction" opacity="0.98" text="Linear regression to burn mid-time">
-                     <toggleGroup>
-                        <ToggleGroup fx:id="toggleGroupRatioCalcMethod" />
-                     </toggleGroup>
-                  </RadioButton>
-                  <RadioButton fx:id="spotAverageRatioCalcRadioButton" mnemonicParsing="false" onAction="#spotAverageRatioCalcRadioButtonAction" opacity="0.98" selected="true" text="Spot Average (time-invariant)" toggleGroup="$toggleGroupRatioCalcMethod" />
-               </children>
-            </HBox>
-            <HBox alignment="CENTER_LEFT" prefHeight="100.0" prefWidth="200.0" GridPane.columnIndex="1" GridPane.rowIndex="4">
-               <children>
-                  <CheckBox fx:id="autoExcludeSpotsCheckBox" mnemonicParsing="false" onAction="#autoExcludeSpotsCheckBoxAction" selected="true" text="Allow Squid to auto-reject spots" />
-                  <Label style="-fx-font-size: 12; -fx-font-weight: bold;" text="Minimum external 1sigma % err for 206Pb/238U:" translateX="20.0" />
-                  <Spinner fx:id="assignedExternalErrUSpinner" editable="true" prefHeight="27.0" prefWidth="70.0" translateX="25.0" />
-                  <Label style="-fx-font-size: 12; -fx-font-weight: bold;" text=" for 208Pb/232Th:" translateX="33.0" />
-                  <Spinner fx:id="assignedExternalErrThSpinner" editable="true" prefHeight="27.0" prefWidth="70.0" translateX="35.0" />
-               </children>
-            </HBox>
-            <HBox alignment="CENTER_LEFT" GridPane.columnIndex="1" GridPane.rowIndex="5">
-               <children>
-                  <Label alignment="CENTER_RIGHT" contentDisplay="TEXT_ONLY" graphicTextGap="0.0" prefHeight="29.0" prefWidth="71.0" style="-fx-font-size: 9.5; -fx-text-fill: red;" text="Def Comm Pb:" textAlignment="RIGHT" wrapText="true" />
-                  <ComboBox fx:id="commonPbModelComboBox" prefHeight="27.0" prefWidth="325.0" promptText="Common Lead Model" />
-                  <Label alignment="CENTER_RIGHT" prefHeight="29.0" prefWidth="65.0" style="-fx-font-size: 10; -fx-text-fill: red;" text="Phys Const:" textAlignment="RIGHT" translateX="10.0" />
-                  <ComboBox fx:id="physConstModelComboBox" prefHeight="27.0" prefWidth="325.0" promptText="Physical Constants Model" translateX="10.0" />
-                  <Button mnemonicParsing="false" onAction="#refreshModelsAction" prefHeight="27.0" prefWidth="65.0" style="-fx-padding: 0 0 0 0;" text="Refresh" translateX="-10.0">
-                     <HBox.margin>
-                        <Insets left="30.0" />
-                     </HBox.margin>
-                  </Button>
-               </children>
-            </HBox>
-            <Label alignment="CENTER_RIGHT" prefHeight="24.0" prefWidth="195.0" style="-fx-font-size: 14;" text="Parameter Models:" GridPane.rowIndex="5" />
-         </children>
-      </GridPane>
       <Label style="-fx-text-fill: #3c77c9; -fx-font-weight: bold;" text="Task Audit:" GridPane.rowIndex="8" />
       <TextArea fx:id="taskAuditTextArea" editable="false" style="-fx-background-color: white; -fx-border-color: black; -fx-font-family: Monospaced; -fx-font-size: 12; -fx-font-weight: bold;" GridPane.columnIndex="1" GridPane.rowIndex="8" />
       <HBox GridPane.columnIndex="1" GridPane.hgrow="ALWAYS">
